+++ packages/api-v4/src/firewalls/types.ts
@@ -51,6 +51,7 @@ export interface FirewallRuleType {
 export interface FirewallDeviceEntity {
   id: number;
   label: null | string;
+  parent_entity: FirewallDeviceEntity | null;
   type: FirewallDeviceEntityType;
   url: string;
 }