+++ packages/manager/src/components/TableRowError/TableRowError.tsx
@@ -1,5 +1,6 @@
 import { ErrorState } from '@linode/ui';
 import * as React from 'react';
+import type { JSX } from 'react';
 
 import { TableCell } from 'src/components/TableCell';
 import { TableRow } from 'src/components/TableRow';