+++ packages/manager/src/features/IAM/Users/UserRoles/UserRoles.tsx
@@ -1,4 +1,4 @@
-import { useAccountUser } from '@linode/queries';
+import { useAccountUser, useUserRoles } from '@linode/queries';
 import {
   CircleProgress,
   ErrorState,
@@ -6,11 +6,10 @@ import {
   Typography,
   useTheme,
 } from '@linode/ui';
+import { useParams } from '@tanstack/react-router';
 import React from 'react';
-import { useParams } from 'react-router-dom';
 
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
-import { useUserRoles } from 'src/queries/iam/iam';
 
 import { AssignedRolesTable } from '../../Shared/AssignedRolesTable/AssignedRolesTable';
 import {
@@ -20,7 +19,7 @@ import {
 import { NoAssignedRoles } from '../../Shared/NoAssignedRoles/NoAssignedRoles';
 
 export const UserRoles = () => {
-  const { username } = useParams<{ username: string }>();
+  const { username } = useParams({ from: '/iam/users/$username' });
   const theme = useTheme();
 
   const {