+++ packages/ui/src/components/Typography/Typography.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Typography } from './Typography';
 
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react-vite';
 
 const meta: Meta<typeof Typography> = {
   component: Typography,