+++ packages/utilities/src/__data__/regionsData.ts
@@ -27,6 +27,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -54,6 +55,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -80,6 +82,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -95,6 +98,7 @@ export const regions: Region[] = [
       'Metadata',
       'Premium Plans',
       'Placement Group',
+      'Maintenance Policy',
     ],
     country: 'us',
     id: 'us-iad',
@@ -109,6 +113,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: ['Linodes'], metrics: ['Linodes'] },
   },
   {
     capabilities: [
@@ -138,6 +143,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -152,6 +158,7 @@ export const regions: Region[] = [
       'Managed Databases',
       'Metadata',
       'Premium Plans',
+      'Maintenance Policy',
     ],
     country: 'us',
     id: 'us-ord',
@@ -166,6 +173,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -194,6 +202,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -222,6 +231,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -250,6 +260,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -277,6 +288,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -304,6 +316,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -332,6 +345,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -360,6 +374,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -387,6 +402,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -416,6 +432,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -443,6 +460,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -472,6 +490,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -499,6 +518,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -524,6 +544,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -554,6 +575,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -584,6 +606,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: ['Linodes'], metrics: [] },
   },
   {
     capabilities: [
@@ -610,6 +633,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -638,6 +662,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: [
@@ -667,6 +692,7 @@ export const regions: Region[] = [
     },
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: ['Linodes', 'Distributed Plans'],
@@ -683,6 +709,7 @@ export const regions: Region[] = [
     },
     site_type: 'distributed',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
   {
     capabilities: ['Linodes'],
@@ -699,5 +726,6 @@ export const regions: Region[] = [
     },
     site_type: 'distributed',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
 ];