+++ packages/utilities/src/factories/regions.ts
@@ -49,6 +49,7 @@ export const regionWithDynamicPricingFactory = Factory.Sync.makeFactory<Region>(
     resolvers: resolverFactory.build(),
     site_type: 'core',
     status: 'ok',
+    monitors: { alerts: [], metrics: [] },
   },
 );