+++ packages/validation/src/linodes.schema.ts
@@ -652,6 +652,11 @@ const CreateVlanInterfaceSchema = object({
   ipam_address: string().nullable(),
 });
 
+const AclpAlertsPayloadSchema = object({
+  system: array().of(number().defined()).required(),
+  user: array().of(number().defined()).required(),
+});
+
 export const CreateVPCInterfaceSchema = object({
   subnet_id: number().required('Subnet is required.'),
   ipv4: object({
@@ -824,5 +829,9 @@ export const CreateLinodeSchema = object({
   firewall_id: number().nullable().notRequired(),
   placement_group: PlacementGroupPayloadSchema.notRequired().default(undefined),
   disk_encryption: DiskEncryptionSchema,
-  maintenance_policy_id: number().notRequired().nullable(),
+  maintenance_policy: string()
+    .oneOf(['linode/migrate', 'linode/power_off_on', undefined])
+    .notRequired()
+    .nullable(),
+  alerts: AclpAlertsPayloadSchema.notRequired().default(undefined),
 });